+++ src/ngx-score.module.ts
@@ -23,4 +23,4 @@ export class NgxScoreModule {
   }
 }
 
-export {createTokenizer, createRenderer, ITokenizer, IRenderer} from './processing'
+export {createTokenizer, createRenderer, ITokenizer, TokenizerType, IRenderer} from './processing'