+++ src/processing/renderer/interfaces.ts
@@ -2,5 +2,5 @@
  * Created by HOFFM59 on 10.05.2017.
  */
 export interface IRenderer {
-  render(data: any);
+  render(data: any): void;
 }