+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 <a name="0.2.0"></a>
 # 0.2.0 (2017-05-11)
 
+* exposes tokenizer and renderer for `TAB` and `musicXML`
+